more apis added, need testing
